Add subject.py
- Move parse_url functions from main.py to subject.py module.
- Create Subject class to hold Course list and methods.
- Add get_valid_course() to main.py.
